Adapted maths functions to use long double to reduce output NaNs due to insufficient precision.

+++ likelihoods.hpp
@@ -14,16 +14,16 @@ std::array<double, 4> computeNucleotideDistribution(const std::vector<Profile>&,
 typedef struct GenomeParameters {
     double heterozygosity;
     double error_rate;
-    std::vector<double> hom_likelihoods;
-    std::vector<double> het_likelihoods;
+    std::vector<long double> hom_likelihoods;
+    std::vector<long double> het_likelihoods;
     GenomeParameters (double h, double e) : heterozygosity{h}, error_rate{e} {}
 } GenomeParameters;
 
 GenomeParameters estimateGenomeParameters(const std::vector<Profile>& profiles, const std::vector<int>& nucleotide_counts);
 
-double profileLikelihoodHomozygous(const Profile& p, const std::array<double, 4>& nucleotide_dist, double p_error);
-double profileLikelihoodHeterozygous(const Profile& p, const std::array<double, 4>& nucleotide_dist, double p_error);
+long double profileLikelihoodHomozygous(const Profile& p, const std::array<double, 4>& nucleotide_dist, double p_error);
+long double profileLikelihoodHeterozygous(const Profile& p, const std::array<double, 4>& nucleotide_dist, double p_error);
 
-double profileLikelihoodHomozygous(const Profile& p, double p_error, int i);
-double profileLikelihoodHeterozygous(const Profile& p, double p_error, int i, int j);
+long double profileLikelihoodHomozygous(const Profile& p, double p_error, int i);
+long double profileLikelihoodHeterozygous(const Profile& p, double p_error, int i, int j);
 #endif
